Prevent leaving the solenoid valve open.
If the solenoid valve is open when the system irrigation configuration is changed to (Off | Schedule | Sensor) the solenoid valve could remain open indefinitely. Close the solenoid valve and write the status as 'Closed'. The subsequent executions of greenhouse.py or the crontab executed scheduled watering opensolenoidtimer.py should return the solenoid valve to the appropriate state (Open | Closed).